Inline IO trait into DrushCommands, remove IOAwareInterface and OutputAwareInterface
The IO trait, IOAwareInterface, and OutputAwareInterface were copied from
Robo in the previous commit but are only used by DrushCommands. Inline
the needed functionality (input/output properties, setInput/setOutput,
currentState/restoreState for SavableState contract) directly into
DrushCommands and delete the intermediate files.

DrushCommands now directly implements the constituent interfaces
(OutputAwareInterface, InputAwareInterface, SavableState) instead of
the removed IOAwareInterface wrapper.

ServiceManager now uses Consolidation\AnnotatedCommand\Output\OutputAwareInterface
directly.
