--- src/Symfony/OutputAwareInterface.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Drush\Symfony;
-
-/**
- * @deprecated Use \Consolidation\AnnotatedCommand\Output\OutputAwareInterface directly
- */
-interface OutputAwareInterface extends \Consolidation\AnnotatedCommand\Output\OutputAwareInterface
-{
-}
